+++ src/crimson/os/alienstore/CMakeLists.txt
@@ -28,7 +28,6 @@ list(APPEND crimson_alien_srcs
   ${PROJECT_SOURCE_DIR}/src/os/bluestore/Allocator.cc
   ${PROJECT_SOURCE_DIR}/src/os/bluestore/AvlAllocator.cc
   ${PROJECT_SOURCE_DIR}/src/os/bluestore/BitmapFreelistManager.cc
-  ${PROJECT_SOURCE_DIR}/src/os/bluestore/BlockDevice.cc
   ${PROJECT_SOURCE_DIR}/src/os/bluestore/BlueFS.cc
   ${PROJECT_SOURCE_DIR}/src/os/bluestore/bluefs_types.cc
   ${PROJECT_SOURCE_DIR}/src/os/bluestore/BlueRocksEnv.cc
@@ -37,23 +36,13 @@ list(APPEND crimson_alien_srcs
   ${PROJECT_SOURCE_DIR}/src/os/bluestore/fastbmap_allocator_impl.cc
   ${PROJECT_SOURCE_DIR}/src/os/bluestore/FreelistManager.cc
   ${PROJECT_SOURCE_DIR}/src/os/bluestore/HybridAllocator.cc
-  ${PROJECT_SOURCE_DIR}/src/os/bluestore/io_uring.cc
   ${PROJECT_SOURCE_DIR}/src/os/bluestore/StupidAllocator.cc
   ${PROJECT_SOURCE_DIR}/src/os/bluestore/BitmapAllocator.cc)
-
-if(HAVE_LIBAIO OR HAVE_POSIXAIO)
-  list(APPEND crimson_alien_srcs
-     ${PROJECT_SOURCE_DIR}/src/os/bluestore/KernelDevice.cc
-     ${PROJECT_SOURCE_DIR}/src/os/bluestore/aio.cc)
-endif()
 
 add_library(crimson-alienstore STATIC ${crimson_alien_srcs}
   $<TARGET_OBJECTS:compressor_objs>
   $<TARGET_OBJECTS:crush_objs>
   $<TARGET_OBJECTS:common_prioritycache_obj>)
-if(HAVE_LIBAIO)
-  target_link_libraries(crimson-alienstore ${AIO_LIBRARIES})
-endif(HAVE_LIBAIO)
 
 target_compile_definitions(crimson-alienstore PRIVATE -DWITH_SEASTAR -DWITH_ALIEN)
 target_include_directories(crimson-alienstore PRIVATE
@@ -64,3 +53,4 @@ target_link_libraries(crimson-alienstore heap_profiler)
 target_link_libraries(crimson-alienstore ${BLKID_LIBRARIES})
 target_link_libraries(crimson-alienstore ${UDEV_LIBRARIES})
 target_link_libraries(crimson-alienstore crimson)
+target_link_libraries(crimson-alienstore blk)
