+++ src/os/CMakeLists.txt
@@ -23,7 +23,6 @@ if(WITH_BLUESTORE)
   list(APPEND libos_srcs
     bluestore/Allocator.cc
     bluestore/BitmapFreelistManager.cc
-    bluestore/BlockDevice.cc
     bluestore/BlueFS.cc
     bluestore/bluefs_types.cc
     bluestore/BlueRocksEnv.cc
@@ -35,19 +34,11 @@ if(WITH_BLUESTORE)
     bluestore/BitmapAllocator.cc
     bluestore/AvlAllocator.cc
     bluestore/HybridAllocator.cc
-    bluestore/io_uring.cc
   )
 endif(WITH_BLUESTORE)
 
-if(HAVE_LIBAIO OR HAVE_POSIXAIO)
-  list(APPEND libos_srcs
-    bluestore/KernelDevice.cc
-    bluestore/aio.cc)
-endif()
-
 if(HAVE_LIBZBC)
   list(APPEND libos_srcs
-    bluestore/HMSMRDevice.cc
     bluestore/ZonedAllocator.cc)
 endif()
 
@@ -55,11 +46,6 @@ if(WITH_FUSE)
   list(APPEND libos_srcs
     FuseStore.cc)
 endif(WITH_FUSE)
-
-if(WITH_BLUESTORE_PMEM)
-  list(APPEND libos_srcs
-    bluestore/PMEMDevice.cc)
-endif()
 
 if(HAVE_LIBXFS)
   list(APPEND libos_srcs
@@ -76,12 +62,8 @@ if(HAVE_LIBZFS)
   list(APPEND libos_srcs $<TARGET_OBJECTS:os_zfs_objs>)
 endif()
 
-if(WITH_SPDK)
-  list(APPEND libos_srcs
-    bluestore/NVMEDevice.cc)
-endif()
-
 add_library(os STATIC ${libos_srcs})
+target_link_libraries(os blk)
 
 target_link_libraries(os heap_profiler kv)
 
@@ -94,25 +76,12 @@ if(WITH_BLUEFS)
   install(TARGETS bluefs DESTINATION lib)
 endif(WITH_BLUEFS)
 
-if(HAVE_LIBAIO)
-  target_link_libraries(os ${AIO_LIBRARIES})
-endif(HAVE_LIBAIO)
-
-if(HAVE_LIBZBC)
-  target_link_libraries(os ${ZBC_LIBRARIES})
-endif()
-
 if(WITH_FUSE)
   target_link_libraries(os FUSE::FUSE)
 endif()
 
 if(HAVE_LIBZFS)
   target_link_libraries(os ${ZFS_LIBRARIES})
-endif()
-
-if(WITH_SPDK)
-  target_link_libraries(os
-    ${SPDK_LIBRARIES})
 endif()
 
 if(WITH_LTTNG)
@@ -133,22 +102,4 @@ if(WITH_BLUESTORE)
     os global)
   install(TARGETS ceph-bluestore-tool
     DESTINATION bin)
-endif()
-
-if(WITH_BLUESTORE_PMEM)
-  target_link_libraries(os pmem::pmem)
-endif()
-
-if(WITH_EVENTTRACE)
-  add_dependencies(os eventtrace_tp)
-endif()
-
-if(WITH_LIBURING)
-  if(WITH_SYSTEM_LIBURING)
-    find_package(uring REQUIRED)
-  else()
-    include(Builduring)
-    build_uring()
-  endif()
-  target_link_libraries(os uring::uring)
 endif()
